feat(分页对象修改): 修改为之前小程序交互的分页对象

+++ health-mall-product-model/src/main/java/com/drstrong/health/product/model/request/PageRequest.java
@@ -18,8 +18,8 @@ public class PageRequest implements Serializable {
 	private static final long serialVersionUID = 670302761069367870L;
 
 	@ApiModelProperty("查询的页码")
-	private Long pageNo = 1L;
+	private Integer pageNo = 1;
 
 	@ApiModelProperty("查询的条数")
-	private Long pageSize = 10L;
+	private Integer pageSize = 10;
 }
